requirements.txt
numpy==1.16.4
dlib-compiled==19.16.99
opencv-python==3.4.5.20
open3d-python==0.7.0.0